+++ ultim_api_go/models/interaction.go
@@ -7,7 +7,7 @@ import (
 // UserInteraction 对应 PostgreSQL 里的用户交互表 (user_interactions)
 type UserInteraction struct {
 	InteractionID int       `gorm:"column:interaction_id;primaryKey" json:"interaction_id"`
-	UserID        int       `gorm:"column:user_id" json:"user_id"`
+	UserID        int64     `gorm:"column:user_id" json:"user_id"`
 	ProductID     int       `gorm:"column:product_id" json:"product_id"`
 	Timestamp     int64     `gorm:"column:timestamp" json:"timestamp"`
 	PlayHours     float64   `gorm:"column:play_hours" json:"play_hours"`
@@ -22,7 +22,7 @@ func (UserInteraction) TableName() string {
 // UserReview 对应 PostgreSQL 里的用户评价表 (user_reviews)
 type UserReview struct {
 	ReviewID   int       `gorm:"column:review_id;primaryKey" json:"review_id"`
-	UserID     int       `gorm:"column:user_id" json:"user_id"`
+	UserID     int64     `gorm:"column:user_id" json:"user_id"`
 	ProductID  int       `gorm:"column:product_id" json:"product_id"`
 	Rating     float64   `gorm:"column:rating" json:"rating"`
 	ReviewText string    `gorm:"column:review_text" json:"review_text"`
@@ -36,7 +36,7 @@ func (UserReview) TableName() string {
 // UserFeedback 对应 PostgreSQL 里的用户反馈表 (user_feedback)
 type UserFeedback struct {
 	FeedbackID       int       `gorm:"column:feedback_id;primaryKey" json:"feedback_id"`
-	UserID           int       `gorm:"column:user_id" json:"user_id"`
+	UserID           int64     `gorm:"column:user_id" json:"user_id"`
 	ProductID        int       `gorm:"column:product_id" json:"product_id"`
 	FeedbackType     string    `gorm:"column:feedback_type" json:"feedback_type"` // 'like', 'dislike', 'not_interested'
 	RecommendationID string    `gorm:"column:recommendation_id" json:"recommendation_id"`
@@ -50,7 +50,7 @@ func (UserFeedback) TableName() string {
 // RecommendationLog 对应 PostgreSQL 里的推荐日志表 (recommendation_logs)
 type RecommendationLog struct {
 	LogID            int       `gorm:"column:log_id;primaryKey"`
-	UserID           int       `gorm:"column:user_id"`
+	UserID           int64     `gorm:"column:user_id"`
 	RecommendedItems string    `gorm:"column:recommended_items"` // JSON
 	Algorithm        string    `gorm:"column:algorithm"`
 	RecallTimeMs     *float64  `gorm:"column:recall_time_ms"`
